Make self-update more robust
In fact, the old method was unsafe. Curl would modify the script
in-place and there was no guarantee that the script is already
in memory, leading to harmless, but unnecessary corruption
at the end of the script.

mv is atomic, so will unlink the old file before writing the new one.