Improve error handling
In most cases, when we run a subshell with our own function
and assign the result to a variable, we also need to check that
the variable is non-empty, since 'die' and 'edo' don't propagate
to the parent shell.

In some cases, non-emptiness is handled in other ways or is not
fatal.