Fix edo in bootstrap-haskell
This seems to have caused occasional weird
side-effects on the following line:

  edo curl https://raw.githubusercontent.com/haskell/ghcup/master/ghcup > "${HOME}"/.ghcup/bin/ghcup
  ...
  edo ghcup install

It would result in:

  ~/.ghcup/bin/ghcup: line 1: $'\E[0': command not found
  ~/.ghcup/bin/ghcup: line 1: 35mcurl: command not found

which seems to be because of printf and the piping command.
It's not entirely clear to me what happens here, but it
seems part of the printf output to stdout where mixed
with the next command.